+++ tests/unit/test_raindrop.py
@@ -102,25 +102,27 @@ def test_is_tagged(
 
 ##############################################################################
 @mark.parametrize(
-    "needle, title, excerpt, note, tags, result",
+    "needle, title, excerpt, note, link, tags, result",
     (
-        ("title", "title", "excerpt", "note", ("tag",), True),
-        ("Title", "title", "excerpt", "note", ("tag",), True),
-        ("excerpt", "title", "excerpt", "note", ("tag",), True),
-        ("Excerpt", "title", "excerpt", "note", ("tag",), True),
-        ("note", "title", "excerpt", "note", ("tag",), True),
-        ("Note", "title", "excerpt", "note", ("tag",), True),
-        ("tag", "title", "excerpt", "note", ("tag",), True),
-        ("Tag", "title", "excerpt", "note", ("tag",), True),
-        ("here", "ishere", "andhere", "alsohere", ("heretoo",), True),
+        ("title", "title", "excerpt", "note", "link", ("tag",), True),
+        ("Title", "title", "excerpt", "note", "link", ("tag",), True),
+        ("excerpt", "title", "excerpt", "note", "link", ("tag",), True),
+        ("Excerpt", "title", "excerpt", "note", "link", ("tag",), True),
+        ("note", "title", "excerpt", "note", "link", ("tag",), True),
+        ("Note", "title", "excerpt", "note", "link", ("tag",), True),
+        ("tag", "title", "excerpt", "note", "link", ("tag",), True),
+        ("Tag", "title", "excerpt", "note", "link", ("tag",), True),
+        ("link", "title", "excerpt", "note", "link", ("tag",), True),
+        ("Link", "title", "excerpt", "note", "link", ("tag",), True),
+        ("here", "ishere", "andhere", "alsohere", "herealso", ("heretoo",), True),
         # Originally I was just smushing all the text-like parts of a
         # Raindrop together, which could result in false positives (actually
         # actual positives but they'd seem false to the average user). This
         # tests that I don't make that mistake again.
-        ("excerpt title", "title", "excerpt", "note", ("tag",), False),
-        ("title note", "title", "excerpt", "note", ("tag",), False),
-        ("note tag", "title", "excerpt", "note", ("tag",), False),
-        ("tag1 tag2", "title", "excerpt", "note", ("tag1", "tag2"), False),
+        ("excerpt title", "title", "excerpt", "note", "link", ("tag",), False),
+        ("title note", "title", "excerpt", "note", "link", ("tag",), False),
+        ("note tag", "title", "excerpt", "note", "link", ("tag",), False),
+        ("tag1 tag2", "title", "excerpt", "note", "link", ("tag1", "tag2"), False),
     ),
 )
 def test_contains(
@@ -128,6 +130,7 @@ def test_contains(
     title: str,
     excerpt: str,
     note: str,
+    link: str,
     tags: tuple[str, ...],
     result: bool,
 ) -> None:
@@ -138,6 +141,7 @@ def test_contains(
             title=title,
             excerpt=excerpt,
             note=note,
+            link=link,
             tags=[Tag(tag) for tag in tags],
         )
     ) is result